+++ onnxruntime/src/session.rs
@@ -724,7 +724,7 @@ unsafe fn get_tensor_dimensions(
     let status = g_ort().GetDimensionsCount.unwrap()(tensor_info_ptr, &mut num_dims);
     status_to_result(status).map_err(OrtError::GetDimensionsCount)?;
 
-    let mut node_dims: Vec<i64> = vec![0; num_dims as usize];
+    let mut node_dims: Vec<i64> = vec![0; num_dims];
     let status = g_ort().GetDimensions.unwrap()(
         tensor_info_ptr,
         node_dims.as_mut_ptr(), // FIXME: UB?
